+++ [Linux]open_in_vscode.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-code "$(dirname "$(realpath "$0")")"