+++ scripts/[Linux]open_in_vscode.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+code "$(cd "$(dirname "$0")/.." && pwd)"